Refactor XML generation using Jsoup
