+++ src/main/java/fr/schiaf/s1000d/model/dmodule/ElementXML.java
@@ -2,6 +2,10 @@ package fr.schiaf.s1000d.model.dmodule;
 
 import java.util.List;
 
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -15,7 +19,33 @@ public abstract class ElementXML {
     List<ElementXML> children;
 
     public abstract String toHtml();
-    public abstract String toS1000DXml();
+    
+    public String toS1000DXml() {
+        Document doc = new Document("");
+        doc.outputSettings().syntax(Document.OutputSettings.Syntax.xml);
+        Element root = doc.appendElement(this.getName());
+        List<ElementXML> attributes = getAttributes();
+        if (attributes != null) {
+            for (ElementXML attribute : attributes) {
+                if (attribute != null && attribute.getChildren() != null && !attribute.getChildren().isEmpty()) {
+                    root.attr(attribute.getName(), attribute.getChildren().get(0).toS1000DXml());
+                }
+            }
+        }
+        for (ElementXML child : this.getChildren()) {
+            root.append(child.toS1000DXml());
+        }
+        return doc.toString();
+    }
 
+    protected void appendChildrenToElement(Element tag) {
+        for (ElementXML child : this.getChildren()) {
+            String childHtml = child.toHtml();
+            Document childDoc = Jsoup.parse(childHtml);
+            for (Element element : childDoc.body().children()) {
+                tag.appendChild(element);
+            }
+        }
+    }
 
 }
