+++ src/main/java/fr/schiaf/s1000d/model/dmodule/tag/IdentAndStatusSection.java
@@ -1,10 +1,8 @@
 package fr.schiaf.s1000d.model.dmodule.tag;
 
 import java.util.LinkedList;
-import java.util.List;
 import java.util.UUID;
 
-import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 
@@ -38,34 +36,8 @@ public class IdentAndStatusSection extends ElementXML {
         Element body = doc.body();
         body.appendElement(HTML_H1).text(S1000D_STATUS);
         Element div = body.appendElement(HTML_DIV);
-
-        for (ElementXML child : this.getChildren()) {
-            String childHtml = child.toHtml();
-            Document childDoc = Jsoup.parse(childHtml);
-            for (Element element : childDoc.body().children()) {
-                div.appendChild(element);
-            }
-        }
-
+        this.appendChildrenToElement(div);
         return doc.toString();
     }
 
-    @Override
-    public String toS1000DXml() {
-        Document doc = new Document("");
-        doc.outputSettings().syntax(Document.OutputSettings.Syntax.xml);
-        Element root = doc.appendElement(this.getName());
-        List<ElementXML> attributes = getAttributes();
-        if (attributes != null) {
-            for (ElementXML attribute : attributes) {
-                if (attribute != null && attribute.getChildren() != null && !attribute.getChildren().isEmpty()) {
-                    root.attr(attribute.getName(), attribute.getChildren().get(0).toS1000DXml());
-                }
-            }
-        }
-        for (ElementXML child : this.getChildren()) {
-            root.append(child.toS1000DXml());
-        }
-        return doc.toString();
-    }
 }
